requirements.txt
Flask>=2.1.0
requests
pandas
google-cloud-secret-manager
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub